+++ data/duke.txt
@@ -1,3 +1,3 @@
-T~done~work
-D~notDone~hw~2023-12-04T12:21
-E~notDone~wedding~2023-12-01T12:00 - 2023-12-02T18:00
+T~done~a
+T~done~b
+T~notDone~c
